for #686: move constant package
