+++ sharding-proxy/src/main/java/io/shardingjdbc/proxy/backend/common/SQLExecuteBackendHandler.java
@@ -29,8 +29,8 @@ import io.shardingjdbc.core.routing.SQLRouteResult;
 import io.shardingjdbc.core.routing.StatementRoutingEngine;
 import io.shardingjdbc.proxy.backend.mysql.MySQLPacketQueryResult;
 import io.shardingjdbc.proxy.config.ShardingRuleRegistry;
-import io.shardingjdbc.proxy.constant.ColumnType;
-import io.shardingjdbc.proxy.constant.StatusFlag;
+import io.shardingjdbc.proxy.transport.mysql.constant.ColumnType;
+import io.shardingjdbc.proxy.transport.mysql.constant.StatusFlag;
 import io.shardingjdbc.proxy.transport.common.packet.DatabaseProtocolPacket;
 import io.shardingjdbc.proxy.transport.mysql.packet.command.query.FieldCountPacket;
 import io.shardingjdbc.proxy.transport.mysql.packet.command.query.TextResultSetRowPacket;
