+++ sharding-proxy/src/main/java/io/shardingjdbc/proxy/transport/mysql/packet/command/quit/ComQuitPacket.java
@@ -17,7 +17,7 @@
 
 package io.shardingjdbc.proxy.transport.mysql.packet.command.quit;
 
-import io.shardingjdbc.proxy.constant.StatusFlag;
+import io.shardingjdbc.proxy.transport.mysql.constant.StatusFlag;
 import io.shardingjdbc.proxy.transport.common.packet.DatabaseProtocolPacket;
 import io.shardingjdbc.proxy.transport.mysql.packet.MySQLPacketPayload;
 import io.shardingjdbc.proxy.transport.mysql.packet.command.CommandPacket;
