+++ sharding-proxy/src/main/java/io/shardingjdbc/proxy/transport/mysql/packet/handshake/HandshakeResponse41Packet.java
@@ -17,7 +17,7 @@
 
 package io.shardingjdbc.proxy.transport.mysql.packet.handshake;
 
-import io.shardingjdbc.proxy.constant.CapabilityFlag;
+import io.shardingjdbc.proxy.transport.mysql.constant.CapabilityFlag;
 import io.shardingjdbc.proxy.transport.mysql.packet.MySQLPacket;
 import io.shardingjdbc.proxy.transport.mysql.packet.MySQLPacketPayload;
 import lombok.Getter;
